MIDI input working, need to debug subsequences
The first subsequence seems correct. It is played when a noteOn is
sent from the EMU keyboard, and waits correctly for a noteOff signalling
the following rest (live player playing the top staff of Study2b3). All
three tracks are playing correctly.
But when a noteOff is sent, the rest is omitted (=bug).