Temporary commit: bug which requires many changes discovered
The running cursor is only updating correctly for the first live noteOn.
The fix will be to remove the cuurent msPositionInScore attribute
altogether from the first message in subsequences, and add an
msPositionInScore attribute to ALL sequences (a single assistant-
only sequence has msPositionInScore=0), then in sequence.tick(),
report (msg.timestamp+sequence.msPositionInScore) as the new position
for the running marker.