Cursor updating correctly in slow live performances
Simplified Sequence.tick(). The cursor still updates correctly
in non-assisted performances. In assisted performeances, the cursor
updates correctly for both noteOn and noteOff if the key is held
down until the subsequence completes. If a noteOff is sent before
the subsequence completes, the cursor gets out of synch with the
audio (the audio is correct).