Beginning to implement performer's continuous controlers
Revised the upper options dialog, including 3 drop-down menues for
setting the continuous controllers which will be associated with the
live performer's 3 continuous controller inputs.
On an EWI, the 3 controllers are aftertouch (breath pressure), pitch
bend and modulation wheel (bite). My E-MU keyboard produces channel
pressure (key pressure), pitch wheel and modulation wheel info.
The keyboard's channel pressure info is treated internally by the
Assistant Performer as if it was really aftertouch info.
The options dialog is now working correctly, and the options are
copied to the jiAPControls options object in the same way as with all
the other options (in apControls.getRuntime().getOptions()), but the
new options do not yet do anything.