Physical pitch-wheel now working.
Connecting the physical pitch-wheel to 'pitch-bend' and the normal
continuous controllers now works (tested with volume and pan), though
this is obviously patch-dependent.
Revised the comments on the upper options panel to describe what I'm
aiming for: NoteOn velocity and pitch will replace the values in the
performer's channel in the score. The performer's controller input will
be added to the controller info in the score.
TODO: physical mod wheel and channel pressure/aftertouch input.
ISSUE: Is the sending of channel pressure working? This is difficult to
test because its patch-dependent. Is this really a 3-byte command? This
needs checking carefully.
KNOWN BUGS: This all needs testing with the Study 3 score...