Overriding with NoteOn pitch and velocity implemented
Both of these now work. The algorithm is as follows:
If one or more of the four options (soloTrackPitch, otherTracksPitch,
soloTrackVelocity, otherTracksVelocity) is set,
a) the lowest piched NoteOn message is found in the first moment in
the subsequence to contain a NoteOnMessage.
b) using this message, the differences between the pitch and velocity
values in the performed message are calculated (pitchDelta and
velocityDelta)
c) pitchDelta and velocityDelta are added to all the appropriate
values in the subsequence. The message in (a) -- in the performer's
track -- thus gets the performed pitch and velocity.
Note that all these searches and calculations are done in real time
when the performer plays a key, and before the subsequence is
actually played. Response may therefore be a bit sluggish in largeLarge subsequences may be a bit sluggish
subsequences.
TODO: test with Study 3 (debug). Write Up in the document 'About
this version of the Assistant Performer'