Assisted performances now start automatically
Simple change: When the 'Start' button in the upper options panel is
clicked, if the MIDI Input is set (i.e. this is an assisted
performance), the 'Go' button is clicked automatically. This simply
avoids a superfluous click, and frustration when the performance
fails to start because one has forgotten to click 'Go'.