Assisted performances can now start anywhere.
An assisted performance can now start anywhere where the start marker
can be set. This includes during a rest in the live performer's track.
If the start marker is set at such a position, a new restSubsequence
is constructed at the given position.
Known Bug: The corresponding situation for end-markers still has to
be completed. Check the situation where both start and end markers
are within a rest in the live performer's track.