Start and end markers working for all tracks.
Both start and end markers can be set anywhere in the soloist's
track. This fixes bug 2 in the previous commit.
The running marker is now updated correctly to the chord or rest
which is playing (while the subsequence is playing) and about to
play (when the subsequence is waiting for a noteOn or noteOff).
Implement the following restrictions on the placing of markers:
It should only be possible to set the start marker to the position
of a performing chord (in any track).
It should only be possible to set the end marker to the position of
a chord or rest in the soloist's track.
Thi smeans that the start marker may move when the track control
changes.