End marker position restrictions implemented.
If an attempt is made to position the end marker during a live
performer's chord, the end marker is positioned at the following
chord or rest in the live performer's part. This applies only to
assisted performances. There are no restrictions in non-assisted
performances.
Known Bug: When both start and end markers are in the same rest.
TODO: Debug very carefully before implementing start marker
restrictions.