End marker bugs fixed.
End markers can now be set in the same live performer's rest as the
start marker.
Tested everything very thoroughly, and ironed out a few other bugs
as well.
TODO: Implement start marker restrictions: it should only be
possible to position them on performing chords (in any track).
This means that they may be moved if the track selection control
is changed. Note that the score is currently given a copy of the
assistant's subsequences (maybe wrongly), and that it probably
needs a copy of the track selection control...