Correction
The score is no longer given a copy of the assistant's subsequences.
TODO: Tell the score which tracks are playing whenever the tracks
control is used. The score should use this info to move the existing
start marker to a performing chord, and to move the start marker to
a performing chord symbol whenever the user resets the start
marker's position.