Start marker completed, performance.now() improved.
svgTracksControl is now given a callback which tells the score when
the number of performing tracks changes. The callback is the score's
performingTracksHaveChanged() function, which updates the position
of the start marker if it is no longer at the position of a
performing chord. The start marker is _always_ at the position of
at least one performing chord.
Added Chris Wilson's code for ensuring that the
window.performance.now() function is always defined. Changed all my
usages of webkitNow() to now(). This may well have been the reason
why the program was not working on Macs.