Coloured coding during assisted performances.
The svgTrackControl now has control of redrawing the whole display.
When the display needs to be redrawn, because the runtime has just
started or the switch between assisted and non-assisted performance
has been made, svgTrackControl.redrawDisplay() is called. This
function updates the control itself, and then calls
score.redrawDisplay().
In assisted performances, the assistants button in the tracksControl
is coloured blue, and cannot be toggled on/off. The assisted track
buttons are light grey, as are their staves in the score. In the
score, the live performers stafflines are black, and its staffname
is bold blue. Staves can have two tracks. If this is the case, the
upper two stafflines denote the upper track (=voice), the lower
three stafflines denote the lower track.
In non-assisted performances, all the buttons in the tracksControl,
and all the stafflines are black.
Stafflines and buttons are still red when 'disabled'.