Beginning to implement MIDI file downloading
Two temporary buttons have been created at the top of the main page.
Button 1 loads a test file from a test directory (temp midi) into
a normal javascript array of numbers.
Button 2 downloads the array as a MIDI file to the user's computer.
When the downloaded file is the same as the original, delete the
temporary buttons, directory and MIDI file, and use the download
function to save performances recorded in the Assistant Performer.