Program logic completed for MIDI file downloading
The 'save' button is now created automatically, when a performance
completes. It responds to mouseover and mouseout events in the same
way as the svgPerformanceControls. It is deleted automatically when
clicked (the file is then downloaded), and when a performance
(both assisted and non-assisted) is started.
All that should now need to be done is to complete the
midiTracksDataToFileArray() function in jiMIDIFile.js.
Bug fix in jiScore.js, line 157: replaced 'chordDef' by 'chordIndex'.> This was preventing the startMarker from being set correctly.