Bug fix: jiScore.js line 157 (checking for chords)
A check is now made for both timeObject.chordIndex and
timeObject.chordDef. The former exists when a timeObject is a chord
which will be retrieved from a palette, the latter when the chord is
defined directly inside an SVG element.
Also: added the images folder (containing the save button images)
to the repositiory.