Bug fixed, another discovered.
Recorded MIDI files now always start with the first message at
time 0.
Known Bug: The timestamps of midiMessages in the score are being
corrupted when the start marker is not at the beginning in a
non-assisted performance. This leads to problems when the 'go'
button is pressed to repeat the performance. This bug is unrelated
to the previous one.