Partial Bug Fix.
Timestamps are no longer being corrupted in non-assisted
performances, so all kinds of non-assisted peformances can now be
repeated and saved. This includes performances with some tracks
turned off, and with the start and end markers at arbitrary
positions.
Known Bug: The assistant still has to be fixed, but it is now clear
how this should be done, so should be no problem tomorrow.
The only file which needs to change now is jiAssistant.js