Beginning to use jiJazzWebMIDIAPI
This commit does not run.
Added js/jiJazzWebMIDIAPI.js which is a drastic revision of Chris
Wilson's shim. Very grateful to him for starting the file, but I
need full control so have made a private copy and am changing it
so that it meets all my requirements. First, I reorganized it so
that JSLint stopped complaining and I could understand what was
going on, then I built it into the Assistant Performer.
Am currently debugging, using abudaan's JazzMidibridge as a crib,
to get it working. (It is NOT yet running).
Have decided to implement a MIDIEvent module before continuing
with the debug. The MIDIEvent module will probably affect what
happens when I continue to debug jiJazzWebMIDIAPI.