Added js/jiMIDIEvent.js
This commit does not run.
jiMIDIEvent exposes constructors for MIDIEvents which are as small
as possible, and are very like the current Web MIDI API object.
Have changed all the files which are affected by the new object,
and they all now pass JSLint. The only file which does NOT pass
JSLint is jiJazzWebMIDIAPI.js, which is going to be affected by
the new jiMIDIEvent.js file. In fact, I think both these files are
going to end up in the Web MIDI API...
jiMIDIEvent.js is where all MIDI constants are currently defined,
and these should be used in jiJazzWebMIDIAPI. In fact, I think
jiMIDIEvent.js should contain all the code not dependent on Jazz.
When the Web MIDI API is running in the browser, it should be
possible to delete jiJazzWebMIDIAPI, and turn parts of
jiMIDIEvent.js into a Javascript library containing those parts
which did not make it into the Web MIDI API.