File reorganisation: JazzWebMIDI and Sequence to lib
I've decided to put all files that can be reused in future MIDI
projects in the lib directory.
Created a new file (a stub) lib/jiJazzWebMIDI.js defining the
webMIDI namespace and having an API identical to the current
Web MIDI API as described at the W3C. Will complete this with the
help of JazzMIDIBridge and WebMIDIAPIShim code.
Other files in lib may end up in a standard library when browsers
implement the Web MIDI API.