Reorganised moment.msPositionInScore and event.timestamps.
This should now be consistent. PositionInScore values are now all
readonly. Timestamps apply only to events, and are not readonly.
Non-assisted performances are working correctly (including
starting later than the beginning of the score), so Sequence.js should
now be correct.
Work still has to be done on assisted performances.