Backup before beginning major timestamp revolution.
In future, Events are not going to have individual timestamps. Instead, each moment will have a (volatile) timestamp in addition to its (constant) msPositionInScore.
This is going to lead to smaller memory use and faster performance because
   a) setTimout() will be called once per moment rather than once per event in tick().
   b) timestamps will be set/reset much less often when starting/ending performances.

Moment.timestamp is always going to be either undefined or an *absolute* DOMHRT time.