Backup commit: Sketch using absolute moment.timestamp
This nearly works, but it might be even better using "relative
timestamps". Try the following TODO:
In score.getSequence() set a moment.msDelay attribute to the
delay between the moment and the previous moment in the
same track. In nextMoment(), the next moment is retrieved
using msPositionInScore. The current absolute timestamp can
be maintained for each track independently using the
(sub)sequence's startNow value.
This algorithm should avoid recalculating timestamps at the
beginning of each playSpan...