Completed timestamp revolution for non-assisted performances.
All files except jiAssistant.js should now work (NEEDS TESTING).
To recapitulate:
Event has been renamed Message.
Messages no longer have timestamps (except temporarily while
converting sequences to Standard MIDI Files), but moments do.
Every moment has a read-only msPositionInScore attribute
which is set when the moments are constructed in
score.createSequence(). (The msPositionInScore attributes take
account of the performance speed option, but this fact can be
ignored in the MIDI lib files.)
At Moment construction time, moment.timestamp attributes are
set to UNDEFINED_TIMESTAMP. The timestamp values are set
properly (to absolute DOMHRTime) in Sequence.nextMoment()
so that they exist both for tick() and in a recorded sequence (if any).
If the performance is being recorded, the moment.timestamps in
the recorded sequence are set relative to the beginning of the
sequence before converting it into a Standard MIDI File. After
conversion, the moment.timestamps in the recording are reset
to UNDEFINED_TIMESTAMP.
If the sequence is not being recorded, moment.timestamps are
reset to UNDEFINED_TIMESTAMP inside tick(), as soon as they are
no longer needed.