Message.getInputEvent comments improved, Sequence.timestamp...
In Sequence.nextMessage(), nextMoment.timestamp is now checked
to see if it has already been set.  If the speed is changing during the
performance, each moment's timestamp is going to be set (relative
to the start of the sequence, using the moment's msPositionInScore
and the current speed) before calling playSpan().