Changed some COMMAND names. Controller options improved.
Changed POLY_AFTERTOUCH to AFTERTOUCH
Changed CHANNEL_AFTERTOUCH to CHANNEL_PRESSURE

Also revised the Assistant.handleController() function.
The Performer's options are now all tested and working,
including the Continuous controller options.

Known Bug: Controller moments are being sent to tracks
which are not actually performing because they currently
inside a rest symbol. Possible solution: Give tracks which
are being recorded an .isResting attribute which is toggled
when a .chordStart or .restStart moment is recorded.