Live performer controls are no longer added to resting tracks.
Renamed Track.addTimestampedMoment to .addLiveScoreMoment
This function now defines and undefines a track.isInChord
attribute which is used to decide whether or not to record any new
live performer's control information.

New function: track.addLivePerformersControlMoment()

Known Bug: The final noteOff message is not being recorded in
assisted tracks. Live performance is correct, but there's a problem
in the recording.