Added timestamp validation to StandardMIDIFile.js
The other midiLib functions should create tracks containing
moments having unique timestamps in ascending order,
but I have (very seldom) experienced exceptions being thrown
showing that this is not always the case.
Probably, the moment.timestamps sometimes get slightly out
of order as a result of awkward thread switching.
This anomalous state should, however, not lead to the loss of
an otherwise good recording, so I have added the function
validateMomentTimestamps(moments) to force a solution:
If a moment's timestamp is less than its predecessor's,
moment.timestamp is set to the predecessor's timestamp.
Equal timestamps should not be produced by the other midiLib code, but this state does not actually lead to the saving of
broken, unplayable Standard MIDI Files.