Improvements for RealTime and SysEx messages.
The Assistant's code has been simplified for the following:
a) it ignores both RealTime and SysEx messages in its input, and
b) assumes that RealTime messages will not interrupt the messages
being received.
The function sequence ToSMF() now throws an exception if an
attempt is made to write a realTime message to a file.
Moved the isRealTimeStatus() utility function to the Constants
namespace.
Updated comments in Constants.js, MIDILibOverview.js etc.