Bug Fix: start marker was not being set correctly in 1-voice scores.
The "set start marker" tool threw an exception in 1-voice scores, because the track indices were not being set properly when there was
only one track.
Also: The scores Study 2c3.1 and Study 3 sketch 1 have been updated.
There is now a link to their respective "About" file at the top of page 1,
and the cursor no longer changes to an I-beam when over SVG text.