New SVG page format
Uploaded new SVG files for Study 2c3.1 and Study 3 sketch 1.
In SVG files, when a chord "use"s a chord definition in the defs section,
the msDuration attribute is the actual duration of the chord in the score
at speed = 1. The msDurations of the basicChords inside the used chord
do not necessarily add up to this value.
When calculating the msDurations of moments, first the total duration of the basicChords is calculated, then the moment msDurations are
set (proportionally to the basicChord msDurations) so that their total
msDuration is equal to the chord's msDuration attribute.
The global speed option simply changes the way the chord's
msDuration attribute is used.
Bug Fix:  system.msDuration is now reset after clicking the Start
button.