Realised silent live performer option
Added a new option check box to the live performer's
options. The performed track can now be silent in live
performances.
In Song Six, I envisage having default sounds in the
soloist's track - which will play back in the default
performance - but that in live performances, the solo
part will be performed live.