Score Infos are now all set in assistantPerformer.html
Deleted ScoreConstants.js.
Controls.js.setScore() now analyses the id of the option selected
in the scoreSelectionControl -- which is defined in
assistantPerformer.html.
This is a much better solution, since each selectable score's
name, nPages and nTracks is now all defined in the same place.
There is no longer any need to change Controls.js when a
score is added or changed in assistantPerformer.html.
Also: selecting "choose a score" in the score selector now
removes the displayed score.