The startRuntimeButton's text now changes.
To complement the disabled status of the button, its
text now changes to one of the values 'not ready',
'Start' or 'Running'. 'Start' is displayed whenever
the button is enabled..
I tried very hard to display a 'wait' cursor, change
the button's text to 'processing...', or display an
image somewhere during the button's onclick() function (i.e. while the MIDI sequence is being
constructed), but without success.