Bug Fix: Live performer's final rest now works correctly
The problem was that the cursor was not advancing
correctly to a rest in the soloist's part when the rest was immediately before the final marker.
The main problem was that the attribute "msPositionInScore" was not being created correctly when cloning restSequences in
Assistant.getPerformedSequences(...).
Also: I simplified things by setting track.toIndex to
the index of the final moment to be played in the
track during a (sub)sequence. This removed the
need to worry about barlines as "duration objects".