Setting startMarker and endMarker revisited
This needed overhauling in the light of the more
advanced features developing in Song Six.  Most
notably, finding systems no longer depends on
their left and right margins. This incidentally
makes selecting the first chord on a staff much
easier - just click in the left margin next to the staff.