Enabled the setting of default performance options
Particular scores are designed to be used with particular
performance options (as set in the main options dialog),
so the defaults can now be set in each score's id string
in the scoreSelector options in assistantPerformer.html.
These defaults can, of course, be overidden before
actually performing the score.
There are comments in both assistantPerformer.html
and Controls.js describing the format of the id string.