Song Six aftertouch is now mapped to volume (buggy)
The volume value sent to the output device  is
64 + half the input aftertouch value. This means that
rests in the controller should be played at (or near)
volume=64.
Bug Fix: The E-MU Modulation wheel now works correctly.
This all works very well, but there is now a KNOWN BUG:
Some of Clytemnestra's syllables are now not sounding.
For example: Her very last "Calls". This does not sound even
when starting the performance there (even though the winds
are sounding)... Is the volume correct when the syllable is
sent?