Simplified assistant.playSpan(). (no more known bugs)
Setting the assistant's performedSequences array in assistant.playSpan(...) has been considerably simplified.
This was possible because the msPositions of the start
and end markers are now much more restricted than
they used to be (some time ago). This simplification was
long overdue.
Also did some refactoring similar to that already done in Sequence.js (Using startMarkerMsPosition and
endMarkerMsPosition make it clearer what these values
actually are.).
The uncontrollability of performance has been much improved by raising the "Retrigger" value in the R2M
(ribbon to MIDI controller) to 80ms (from 0!).