Reorganized volume and aftertouch/channel pressure
The Assistant performer no longer looks for volume
messages in SVG-MIDI files.
(None of the existing scores contains volume control
info, and I am about to remove them from the
SVG-MIDI specification (and Moritz' score writing
software.)
In performances without a live performer: the AP sets
each track's volume to 125 before the performance
begins. No further volume messages are sent during
the performance.
In performances _with_ a live performer:
There is a new option "po.pressure.minimum=" that
can be set with the other performance options in
assistantPerformer.html. If pressure is mapped to
volume, this is the volume value to which all tracks
are initialised before the performance begins. If the
option is not set the volumes are initialised to 127.
If the option is mapped to some other control,
those controls are initialized to that value, and
volume is initialised to 127.
During the performance, pressure values (aftertouch
or channel pressure) are sent to the mapped
controller on all performing tracks, whereby the
pressure values (range 0..127) are mapped to the
range po.pressure.minimum..127.