Minimum pressure setting in main options panel
The minimum pressure setting, set by the value in the
score's id string in assistantPerformer.html, is now
available in the main options when either of the
pressure checkboxes is checked. The performer can
thus override the default value in the html file.
An alert is displayed if the setting is illegal when the
start button is pressed. (The same is now true for the
speed factor.)
Live performances are no longer set 'playing' when
the start button is pressed.