Completed performer's pressure/volume option.
Changes to assistantPerformer.html, Controls.js and
Assistant.js. Did some refactoring in all three files.

In Controls.js: In assisted performances, options now
has a "runtimeOptions" attribute. This object has
trackMinVolumes (an array of integers in the range
0..127, one value per track), and trackScales (an array
of floats in the range 0.0..1.0, one value per track)
attributes. If the performer is sending volume control
messages, these arrays are used with the performer's
input control value at runtime to calculate the values
of the volume messages actually sent to each track.