Beautified: score options are now set just once.
A score's default options are now set once
(in Controls.js setScore() ), at the same time as the
SVG pages are loaded. This allowed some functions
to be moved inside others, thus reducing the scope
in which they are available.
Added a temporary track.maxVolumes option to
Study2c3.1 to test that this works in non-assisted
performances (it does).