Performer's minimum volume option revisited.
The "minPressureDiv" has been renamed "minVolumeDiv",
and all names that depend on that option in Controls.js
have been refactored to use Volume instead of Pressure.
Any of the three control mapping selectors can be mapped
to volume, and the minVolumeDiv is therefore visible/hidden depending on whether any of the three selectors
is both visible and mapped to volume.