Changes for running from a local server on the user's computer.
If the assistantPerformer directory is copied to a local server,
assistantPerformer.html will now run correctly from there in
Chrome if Chrome is started with its --disable-web-security
flag set.
Full URLs for score pages are set with respect to the full URL
of the start page (assistantPerformer.html), so does not really
matter where the start page starts -- on the web, on a Mac, or
on a Windows machine.