The music font is now included in the assistantPerformer directory.
This allows the font (Clicht) to be loaded relatively to the score,
which means that it can load both on the web and from a local
hard drive.
The score generation code has been modified accordingly (in
Moritz), and all the score pages have been recompiled. They
now address the font stylesheet relatively.