Added defaultPerformersOptions.assistantsTrackIndex
Changes to the score of Song Six (in Moritz):
added three more (as yet empty) tracks for furies, and moved the
R2M control track down to track 5 (index 4 in the performance
options track selector). Song Six now has 9 pages, but this may
increase as the furies get more ledgerlines.
This change in the score meant that I needed to be able to set
the default control track here in the Assistant Performer.