Minimum pressure option completed (!)
The minimum pressure option provides a lower limit to the
pressure values (Aftertouch or ChannelPressure) received
by the assistant from the live performer.
This ensures that the assistant's volumes do not get set to 0
when the performer begins a rest.
This option had only been _partially_realized... It now works!