Send only NOTE_OFFs when silently finishing sequences.
Sequence.finishSilently() is called when the live
performer plays a rest before the current (sub)sequence
is finished. Sending just the NOTE_OFFs (and NOTE_ONs
with zero velocity) from the remaining part of the
sequence ensures that nothing more is heard of it.
This assumes that all LocalMidiDurationDefs are
responsible for setting their own context (Patch etc.)