NoteOffs are now only sent when necessary.
Many redundant noteOff messages are no longer
sent at the ends of MIDIChords.
The new code for sending noteOffs when they ARE
necessary at the ends of MIDIChords is called by
none of my current scores, so it has not been tested.
Such (untested) noteOffs are necessary at the ends
of MIDIChords having a chordOff, but containing
basicMidiChords that have no chordOff.

Also in this commit:
The score of Song Six and assistantPerformer.html
have been updated for new system breaks. There are
now 13 pages instead of 14.