Progress -- but buggy
Workers are still not defined, and there are bugs
which have to be fixed first.
This commit contains some special comments
in Tracks.js and Sequence.js (beginning with
"// *** ") as notes for how the workers might
eventually be used.
Currently debugging just with a single channel
(Song Six, wind 2). This should be working before
before beginning to write any real worker code.