Added ChordDef.repeatMoments attribute.
And simplified the code again(!)
If ChordDef.repeatMoments is undefined, the
default is that moments in a chord repeat
endlessly. If they dont repeat, the next moment
is null after they have all been played once.
The current chord in each track is updated
when the following chord's msPositionInScore
is the position reached in the performance.
This is working for simple playback of Song Six.