Renamed Performer.js-->Player.js
And renamed namespace "performer" to "player".
The "player" is like a "MIDI file player" which can
be influenced by a live performer. So I want to
reserve the word 'performer' to mean a real
person. The docs will usually refer to a
"live performer".
Also: renamed score.createSequence to
score.initializePlayer and moved the call to
player.init(...) inside that function.