Performer.js renamed to MonoInput.js
When complete, MonoInput.js will define the
monoInput namespace defining an event
handler for a mono input device such as an
R2M or EWI.
Also
1. Added the file PolyInput.js. This currently
just contains stubs, but will eventually define
the polyInput namespace which handles each
key of a keyboard independently. This can be
used to realize things like prepared pianos.
2. Updated assistantPerformer.html, Controls.js
and Player.js for the new, projected namespaces.