Completed new assistantPerformer.html
This page now has a new performer options
div including all the monoPerformerOptions
as working controls. These have to be set
when the mono options load. They all call a
doControl() function in the monoInput
namespace. This function is currently just a stub.
assistantPerformer.html now also includes an
empty div stub for the polyPerformerOptions.
The old options handling needs to be removed
from Controls.js.