Revised the performers options.
This version introduces midi event handlers as
separate files having the same public interface.
There are currently three of these. The play()
function in the appropriate file is called when
the user clicks the main start button to display
the score:
1. ScorePlayer. js plays a score without waiting for
input from a midi input device.
2. Mono1.js is currently a copy of ScorePlayer.js
but includes a handleMIDIInputEvent function
in its code and public interface. The play()
function needs to be changed so that it waits for
midi input to arrive at handleMIDIInputEvent.
3. Keyboard1.js is a copy of Mono1. I'm intending
to write a "prepared piano" handler here.
These handlers currently contain lots of duplicated
code. I originally intended that they would be
derived from a common prototype, but this turned
out to be impractical. I suspect that they may
eventually use functions from a common helper
namespace.