Preparing not to use constructed subsequences.
Because it would be difficult to organise the
borders with midiChords repeating while waiting
for noteOff events in assisted performances.
Stage 1 complete: unassisted performances are
working after having made apropriate changes to
sequence.init() and sequence.play().
TODO: Mono1 and the bridges between the live
performer's noteOff/noteOns. I think I need a new
sequence state: continuing, which is started when
a sequence range reports that it has come to its
end, and a noteOn has not yet arrived to start the
next range of the performance.