ScoreMsPosition bug fixed in Sequence.js
Score synchronisation is now only related
to the msPositions in the score. It no longer
has anything to do with passing time! :-)
This becomes important when speeds are
changing in assisted performances.
Line 264, using performance.now(), was the
main culprit.
Score performances have been tested and are
working.
TODO: test and debug assisted performances.