Mono1 now working correctly in Song Six.
NoteOn and NoteOff handlers now work as they
should in Song Six. Has been tested with different
positions for the start and end markers.
Known bug: endMarkerPositions should only be
possible where the live performer has a chord or
rest. Currently, I think it is possible for them to end
up at the end of a system where this is not the case.