Bug fix: mono1 now working better. Known bugs.
Bug 1: Clytemnestra's notes are set to repeat in the
score. This is wrong (the score needs changing),
but the notes should move on anyway when the
cursor moves on. This needs debugging for both
assisted and non-assisted performances.
Bug 2: Debug an assisted performance of Study 2,
staff 2 alone. This contains an irregular sequence
of chords and rests -- not just alternating chords
and rests as in Song Six.