Implementation of chord looping in Sequence.js
Non-assisted performances of both Study 2 and
Song Six have been tested, also with stopping,
pausing, restarting with and without new start and
end markers etc.
Looping has been prepared, but is off by default
in Sequence.js. It should be possible to turn it on
during live performances using the
sequence.setLooping(bool) function -- for
example when waiting for a noteOn or noteOff.
This has not yet been attempted...