Chord looping completed, tested and working.
Unassisted performances are also still working
as before.
In live performances, midiChords which
do not have a Repeat=0 attribute in the score,
now repeat when the live performer is holding
a key down and all the tracks are playing their
final midiObject in the span. This has been
tested, and is working.
TODO: The default in the score should be for
midiChords NOT to repeat. They should only do
so if they have a Repeat = 1 attribute (and the live
performer is holding a key down while they are
playing as the final midiObject in the track in the
span.