Changed for abudaan's version of WebMIDIAPI shim
This version of the shim mimiks Chrome's current
behaviour, so that the code in Controls.js can be the
same for Chrome with its WebMIDI flag set and
Firefox using Jazz. The code has therefore been
changed accordingly.
This (current development) version of the AP is
loading its devices correctly in both Chrome and FF,
but it does not yet, support the new SVG-MIDI file
format, so scores don't yet play.
Also changed the corresponding parts of the *public*
Assistant Performer on my website, so that that *is*
now playing correctly.