Major Commit (Restarting Git repository 2015)
There have been major changes to many files.
In particular, the initial options dialog and the
tracksControl have been completely revised.
Also, account is now taken of input and output
voices/tracks.
Currently:
1. There are no known bugs when performing
scores that do not contain input voices.
2. InputChord.js, which contains the InputChord
constructor, needs to be completed (that's the
next major thing that has to be done).
3. Keyboard1.js, which will contain the input event
handler for the "prepared keyboard" algorithm,
needs to be completed.