InputChord code is now complete and working.
Seq* has been renamed to Trk* in Moritz, in the
score file format, and in this project. A Trk is a
subsection of a Track, not a subsection of a
Sequence. All the scores have therefore been
recompiled.
Each InputNote currently has a trks attribute
containing the trks that are triggered when the
key is pressed. Should these be a Seq?