Added Seq.js -- defines Seq objects
A Seq is like a Sequence, but has a different
public interface. Seqs and Sequences will
probably share some code in future...
Each inputNote has a seq that will be triggered
when the inputNote is activated.
Seqs are going to run in background worker
threads and react to incoming messages.
Seq.js is currently just a stub. Some initialization
still needs to be done in the constructor, and public
message handling functions need to be added.