Running Marker now moves correctly
"Study 3 sketch 2 with input" can now be
"played", with the running marker updating
correctly when the correct pitches are
performed on the keyboard.
Input options are traced, but don't yet have
any effect.
Seqs are not yet played, but there is a stub
function for doing that.
Some refactoring has been done.