Revised the setting of start and end markers
This was a major change, but it has been
tested thoroughly, and start and end
markers are now being set correctly.
The behaviour has changed slightly, but
some old code has been removed and the
new code is simpler.
Score now maintains its own internal
isLivePerformance and trackIsOnArray
variables. score.trackIsOnArray is
initialized to all tracks on when the score
is loaded, and reset when the trackControl
calls score.refreshDisplay.